requirements.txt
pyfiglet
python-nmap
